version 1.2
Change the game structure
