+++ food.py
@@ -1,0 +1,7 @@
+import turtle
+
+class Food(turtle.Turtle):
+    def __init__(self):
+        super().__init__()
+        self.penup()
+        self.shape('apple_.gif')
