+++ high_score.txt
@@ -1,1 +1,1 @@
-2
+27
